upgrade kotlin to 2.3.0 and set Java 25 as new baseline (#205)
* Refactor: Simplify type conversions and update dependencies across the project

* Refactor: Consolidate `sceneParameter` logic and update Java toolchain to version 25

* Refactor: Increase JVM memory limits, enforce `internal` visibility for `sceneParameter`, and downgrade Java toolchain to version 17

* Refactor: Add Gradle profiling and coverage controls to run configurations

* Refactor: Update logging, testing, and plugin dependencies in `libs.versions.toml`

* Refactor: Update OS matrix to use latest macOS version in GitHub Actions workflow

* Refactor GPUBuffer functionality: remove deprecated mapping methods and upgrade usage handling to GPUBufferUsage flags.

* Remove platform-specific `ArrayBuffer` implementations in favor of unified handling. Cleanup deprecated `writeTexture` and `writeBuffer` methods.

* Refactor buffer and texture usage handling: simplify `writeBuffer` and `writeTexture` logic, unify GPU usage flags, and remove deprecated methods.

* Remove deprecated `writeInto` methods and replace `ArrayBuffer.from` with `ArrayBuffer.of`. Cleanup unused imports.

* Update `webgpu-ktypes` version to stable `0.0.9`.

* Remove commented-out `publishLibraryVariants` references from Gradle files.

+++ wgpu4k/src/commonNativeMain/kotlin/mapper/TextureViewDescriptor.kt
@@ -8,13 +8,13 @@ internal fun MemoryAllocator.map(input: GPUTextureViewDescriptor) = WGPUTextureV
     .also { output ->
 
         map(input.label, output.label)
-        input.format?.let { output.format = it.value.toUInt() }
-        input.dimension?.let { output.dimension = it.value.toUInt() }
-        output.aspect = input.aspect.value.toUInt()
-        output.baseMipLevel = input.baseMipLevel.toUInt()
+        input.format?.let { output.format = it.value }
+        input.dimension?.let { output.dimension = it.value }
+        output.aspect = input.aspect.value
+        output.baseMipLevel = input.baseMipLevel
         // @see https://developer.mozilla.org/en-US/docs/Web/API/GPUTexture/createView#miplevelcount
         output.mipLevelCount = input.mipLevelCount ?: 1u
-        output.baseArrayLayer = input.baseArrayLayer.toUInt()
+        output.baseArrayLayer = input.baseArrayLayer
         // @see https://developer.mozilla.org/en-US/docs/Web/API/GPUTexture/createView#arraylayercount
         output.arrayLayerCount = input.arrayLayerCount ?: 1u
     }
